refactor: extract backend podcast mock handler to separate file
